section 5  done

+++ src/AuctionService/Controllers/AuctionController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -54,11 +55,12 @@ public class AuctionsContoller : ControllerBase
     }
 
     [HttpPost]
+    [Authorize]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = _mapper.Map<Auction>(auctionDto);
-        //TODO
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
 
         _context.Auctions.Add(auction);
 
@@ -76,6 +78,7 @@ public class AuctionsContoller : ControllerBase
 
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
     {
@@ -84,7 +87,7 @@ public class AuctionsContoller : ControllerBase
         if(auction == null)
             return NotFound();
 
-        //TODO
+       if(auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -101,6 +104,7 @@ public class AuctionsContoller : ControllerBase
         return Ok();
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id)
     {
@@ -109,7 +113,7 @@ public class AuctionsContoller : ControllerBase
         if(auction == null)
             return NotFound();
 
-        //TODO
+        if(auction.Seller != User.Identity.Name)  return Forbid();
 
         _context.Auctions.Remove(auction);
